requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile requirements.in
#
blinker==1.9.0
    # via flask
click==8.1.8
    # via flask
colorama==0.4.6
    # via click
dnspython==2.7.0
    # via
    #   -r requirements.in
    #   pymongo
flask==3.1.0
    # via
    #   -r requirements.in
    #   flask-cors
    #   flask-sqlalchemy
flask-cors==5.0.1
    # via -r requirements.in
flask-sqlalchemy==3.1.1
    # via -r requirements.in
greenlet==3.1.1
    # via sqlalchemy
itsdangerous==2.2.0
    # via flask
jinja2==3.1.6
    # via flask
markupsafe==3.0.2
    # via
    #   jinja2
    #   werkzeug
pymongo==4.11.3
    # via -r requirements.in
python-dotenv==1.1.0
    # via -r requirements.in
sqlalchemy==2.0.39
    # via flask-sqlalchemy
typing-extensions==4.13.0
    # via sqlalchemy
werkzeug==3.1.3
    # via
    #   flask
    #   flask-cors
